server端适配调整

+++ canal-admin/canal-admin-server/src/main/java/com/alibaba/otter/canal/admin/service/impl/CanalClusterServiceImpl.java
@@ -2,6 +2,8 @@ package com.alibaba.otter.canal.admin.service.impl;
 
 import java.util.List;
 
+import com.alibaba.otter.canal.admin.common.exception.ServiceException;
+import com.alibaba.otter.canal.admin.model.NodeServer;
 import org.springframework.stereotype.Service;
 
 import com.alibaba.otter.canal.admin.model.CanalCluster;
@@ -25,6 +27,12 @@ public class CanalClusterServiceImpl implements CanalClusterServic {
     }
 
     public void delete(Long id) {
+        // 判断集群下是否存在server信息
+        int serverCnt = NodeServer.find.query().where().eq("clusterId", id).findCount();
+        if (serverCnt > 0) {
+            throw new ServiceException("Servers exist, delete failed");
+        }
+
         CanalCluster canalCluster = CanalCluster.find.byId(id);
         if (canalCluster != null) {
             canalCluster.delete();
